// typechain-types/contracts/merkleVesting/VTVLMerkelVesting.ts
export type ClaimInputStruct = {
  startTimestamp: BigNumberish;
  endTimestamp: BigNumberish;
  cliffReleaseTimestamp: BigNumberish;
  releaseIntervalSecs: BigNumberish;
  scheduleIndex: BigNumberish;
  linearVestAmount: BigNumberish;
  cliffAmount: BigNumberish;
  recipient: string;
};